rename plugin/ to fixer/
