+++ src/fixer/fix/apply-auto-fixes.test.ts
