+++ src/fixer/fix/apply-auto-fixes.ts
