+++ src/fixer/fix/apply-suggestions.test.ts
