+++ src/fixer/fix/make-fixable-and-fix.test.ts
