+++ src/fixer/prefer-addition-shorthand-rule.ts
