+++ src/fixer/report-translator.ts
